styled editors and output
